refactor(25265): refactor main icons
