+++ hivemq-edge/src/frontend/src/locales/en/components.json
@@ -1,6 +1,9 @@
 {
-  "topic": {
-    "iconLabel": "Topic"
+  "iconLabel": {
+    "topic": "Topic",
+    "tag": "Tag",
+    "client": "Client",
+    "workspace": "Workspace"
   },
   "pagination": {
     "ariaLabel": "Pagination",
