+++ hivemq-edge/src/frontend/src/modules/Dashboard/hooks/useGetNavItems.tsx
@@ -8,7 +8,7 @@ import { GoLinkExternal } from 'react-icons/go'
 import { MdOutlineEventNote, MdPolicy } from 'react-icons/md'
 
 import { useGetConfiguration } from '@/api/hooks/useFrontendServices/useGetConfiguration.ts'
-import WorkspaceIcon from '@/components/Icons/WorkspaceIcon.tsx'
+import { WorkspaceIcon } from '@/components/Icons/TopicIcon.tsx'
 
 import config from '@/config'
